+tests

+++ src/main/java/Main.java
@@ -4,6 +4,6 @@ public class Main {
     public static void main(String[] args) {
         FunctionTanX calculator = new FunctionTanX();
         for (int i = 1; i <= 20; i++)
-            System.out.println(i + " : " + calculator.calculate(i).floatValue());
+            System.out.println(i + " : " + calculator.calculate(i));
     }
 }
